Reorg: Move test utils
