+++ src/test/java/com/ginsberg/gatherers4j/BigDecimalProductGathererTest.java
@@ -17,6 +17,7 @@
 package com.ginsberg.gatherers4j;
 
 import com.ginsberg.gatherers4j.dto.WithOriginal;
+import com.ginsberg.gatherers4j.util.TestValueHolder;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
